feat: addToCalendar function
I needed the ability to use the wonderful "add to calendar" _action_
without using the same HTML-based setup & initialization.

This renames `atcb_open` to `atcb_addToCalendar` and reworks it to
generate the dropdown list and background overlay _only when clicked_.
This has no discernible performance impact, while making it possible to
also use the function directly, without needing `atcb_init`.

In the NPM package, the prefix is removed, so that NPM users can:

    import { addToCalendar } from 'add-to-calendar-button'

This also positions the dropdown list below the relevant button using
absolute positioning while putting it elsewhere in the DOM. This
required some CSS tweaks. It will also help avoid z-index styling
issues, which can be a problem when putting dropdowns directly adjacent
to the trigger within the DOM.

If someone calls `addToCalendar` without specifying a second argument,
the dropdown list is shown as a modal.

Finally, the new format required getting rid of the implicit `this`
binding. Instead, arrow functions have been used throughout.